Change column amount to be a float

+++ db/migrate/20221115085751_add_amount.rb
@@ -1,5 +1,5 @@
 class AddAmount < ActiveRecord::Migration[7.0]
   def change
-    add_column :transactions, :amount, :integer
+    add_column :transactions, :amount, :float
   end
 end
